+++ src/tensor/tensor.cpp
@@ -165,7 +165,8 @@ void Tensor::debug() const {
 
 bool Tensor::isContiguous() const {
     ptrdiff_t expected_stride = 1;
-    for(size_t i = ndim() - 1; i >= 0; i--){
+    
+    for(int i = static_cast<int>(ndim()) - 1; i >= 0; i--){
         if(_meta.strides[i] != expected_stride){
             return false;
         }
@@ -213,7 +214,7 @@ tensor_t Tensor::view(const std::vector<size_t> &shape) const {
     size_t ndim_ = shape.size();
     new_meta.strides.resize(ndim_);
     size_t stride = 1;
-    for(size_t i = ndim_ - 1; i >= 0; i--){
+    for(int i = static_cast<int>(ndim_) - 1; i >= 0; i--){
         new_meta.strides[i] = stride;
         stride *= shape[i];
     }
